Server/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
email-validator
python-dotenv
pymysql
python-jose[cryptography]
cryptography